Refactored some common objects into orders
